+++ mvs_eland_tool/mvs_eland_tool.py
@@ -31,25 +31,14 @@ import logging
 import os
 
 # Loading all child functions
-try:
-    # for tests
-    from .src.A0_initialization import initializing
-    from .src.A1_csv_to_json import DataInputFromCsv
-    from .src.B0_data_input_json import data_input
-    from .src.C0_data_processing import data_processing
-    from .src.D0_modelling_and_optimization import modelling
-    from .src.E0_evaluation import evaluation
-    from .src.F0_output import output_processing
 
-except ModuleNotFoundError:
-    # for terminal execution
-    from src.A0_initialization import initializing
-    from src.A1_csv_to_json import DataInputFromCsv
-    from src.B0_data_input_json import data_input
-    from src.C0_data_processing import data_processing
-    from src.D0_modelling_and_optimization import modelling
-    from src.E0_evaluation import evaluation
-    from src.F0_output import output_processing
+import src.A0_initialization as initializing
+import src.A1_csv_to_json as load_data_from_csv
+import src.B0_data_input_json as data_input
+import src.C0_data_processing as data_processing
+import src.D0_modelling_and_optimization as modelling
+import src.E0_evaluation as evaluation
+import src.F0_output as output_processing
 
 
 def main(**kwargs):
@@ -78,14 +67,14 @@ def main(**kwargs):
     # Read all inputs
     #    print("")
     #    # todo: is user input completely used?
-    #    dict_values = data_input.get(user_input)
+    #    dict_values = data_input.load_json(user_input)
 
     logging.debug("Accessing script: A1_csv_to_json")
-    path_to_json_from_csv = DataInputFromCsv.create_input_json()
+    path_to_json_from_csv = load_data_from_csv.create_input_json()
     user_input.update({"path_input_file": path_to_json_from_csv})
 
     logging.debug("Accessing script: B0_data_input_json")
-    dict_values = data_input.get(user_input)
+    dict_values = data_input.load_json(user_input)
 
     print("")
     logging.debug("Accessing script: C0_data_processing")
@@ -108,8 +97,8 @@ def main(**kwargs):
                                   filename=dict_values['simulation_settings']['oemof_file_name'])
             results_main = model.results['main']
             results_meta = model.results['meta']
-            
-    """
+
+    # """
     print("")
     logging.debug("Accessing script: E0_evaluation")
     evaluation.evaluate_dict(dict_values, results_main, results_meta)