+++ tests/constants.py
@@ -1,0 +1,3 @@
+import os
+
+REPO_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))